==> Genebass_VEPs/2-1_from_Genebass_to_VESPA.ipynb <==
## README

make input files and command file(bash) for running VESPA

Input Files
Required: A single FASTA file containing all your wildtype sequences (note: this file can contain any number of sequences).

Optional: If you are only interested in a subset of possible mutations (specific mutations), you can add -m mutations.txt to the code line in Quickstart (note: per default all mutations are considered). Click here to head to the file format explanation of mutations.txt.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import r2_score
import re
import requests
import json


## Fetch FASTA file altogether into a file

def get_fasta_merge(uniprot_id):
    # uniprot_id = "P07550"
    url = f"https://www.uniprot.org/uniprot/{uniprot_id}.fasta"

    response = requests.get(url)
    if response.ok:
        fasta_data = response.text
        with open("sequence.fasta", "a") as f:
            f.write(fasta_data)
    else:
        print(f"Error retrieving FASTA sequence: {response.status_code} {response.reason}")

# Load the array from the text file
unq_uniid_list = np.loadtxt("unq_uniprot_id_68_forGB.txt", dtype='str')
unq_uniid_list

# put all fasta sequence into one fasta file
for i in range(len(unq_uniid_list)):
    get_fasta_merge(unq_uniid_list[i])
    

## Make mutation list(Missense mutation!!!!) for all target proteins into a txt file

VESPA! mutation position = position - 1!!!

# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
gene_list_raw.rename(columns={'gene':'SYMBOL'},inplace=True)
gene_list_raw
# uni_list = gene_list_raw['Uniprot'].tolist()


VEPwithGB_icd10 = pd.read_csv('VEPwithGB_icd10_0323.csv', index_col = None)
VEPwithGB_icd10

VEPwithGB_icd10['mutation_from'] = VEPwithGB_icd10['Amino_acids'].str.split('/',expand=True)[0]
VEPwithGB_icd10['mutation_to'] = VEPwithGB_icd10['Amino_acids'].str.split('/',expand=True)[1]
VEPwithGB_icd10

VEPwithGB_icd10_uniid = pd.merge(VEPwithGB_icd10,gene_list_raw[['protein','SYMBOL','Uniprot']],on='SYMBOL',how='left')
VEPwithGB_icd10_uniid

unq_VEPwithGB_icd10 = VEPwithGB_icd10_uniid.drop_duplicates(subset=['SYMBOL','Protein_position','mutation_from','mutation_to'], keep='first')
unq_VEPwithGB_icd10 = unq_VEPwithGB_icd10.reset_index()
unq_VEPwithGB_icd10

from above, there are in total 15,491 missense mutation


for i in range(unq_VEPwithGB_icd10.shape[0]):
    # get the uniprot id according to entryname and convert the Series to a string with index excluded
    uniprot_id = unq_VEPwithGB_icd10.iloc[i,67]
    position = unq_VEPwithGB_icd10.iloc[i,3]-1
    mut_info = unq_VEPwithGB_icd10.iloc[i,64]+str(position)+unq_VEPwithGB_icd10.iloc[i,65]
    with open("mutations_gb.txt", 'a') as file:
        # write a new line to the file
        file.write(uniprot_id+"_"+mut_info+"\n")
    

## Process the output of VESPAI and output as csv file

import json
import pandas as pd
import os

# read in the map.json file
with open('output/map.json', 'r') as f:
    id_map = json.load(f)

# create an empty dataframe to store the combined data
combined_df = pd.DataFrame(columns=['Uniprot_ID', 'SYMBOL', 'Mutant', 'VESPAl'])

# loop over each file in the output folder
for filename in os.listdir('output'):
    if filename.endswith('.csv'):
        # get the uniprot id for this file from the filename
        id_num = int(filename.split('.')[0])
        uniprot_id = id_map[str(id_num)]

        symbol = gene_list_raw.loc[gene_list_raw['Uniprot']== uniprot_id,'SYMBOL'].to_string(index=False)
        
        # read in the csv file and split the information on ';'
        csv_data = pd.read_csv(os.path.join('output', filename), delimiter=';')
        
        # add the uniprot id column and concatenate with the combined dataframe
        csv_data['Uniprot_ID'] = uniprot_id
        csv_data['SYMBOL'] = symbol
        combined_df = pd.concat([combined_df, csv_data], ignore_index=True)

# print the resulting dataframe
print(combined_df)


# extract amino acid, position, and end amino acid from 'Mutant' column
combined_df[['mutation_from', 'mutation_pos', 'mutation_to']] = combined_df['Mutant'].str.extract(r'(\D)(\d+)(\D)')

# convert 'Position' column to integer type and plus 1 for further merge with GPCRdb data
combined_df['mutation_pos'] = combined_df['mutation_pos'].astype(int)+1
combined_df = combined_df.drop('Mutant', axis=1)
combined_df

combined_df.to_csv("GB_vespa_df.csv",index=False)
