==> Genebass_VEPs/2-2_VESPA_score_Genebass_pb.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import r2_score
import re
import requests
import json


vespa_df = pd.read_csv("GB_vespa_df.csv")
vespa_df = vespa_df.rename(columns={'mutation_pos':'Protein_position'})
vespa_df

icd10_df = pd.read_csv('VEPwithGB_icd10_0323.csv')
icd10_df

icd10_df['mutation_from'] = icd10_df['Amino_acids'].str.split('/',expand=True)[0]
icd10_df['mutation_to'] = icd10_df['Amino_acids'].str.split('/',expand=True)[1]


icd10_df

# merge two dataframes
gb_with_vespa  = pd.merge(icd10_df,vespa_df,on=['Protein_position','mutation_from','mutation_to','SYMBOL'])
gb_with_vespa

gb_with_vespa.to_csv('Genebass_with_VESPAl_df.csv',index=False)

## [not used]classification performance(seperately)

from sklearn.metrics import accuracy_score, recall_score, precision_score, f1_score, classification_report, roc_curve, auc, confusion_matrix

gpcr_vespa['VESPAl']

# VESPAl into two classifications: >=0.5(positive) <0.5(negative)
gpcr_vespa['VESPAl_group'] = gpcr_vespa['VESPAl'].apply(lambda x: 1 if x >= 0.5 else 0)
gpcr_vespa['foldchange_group'] = gpcr_vespa['exp_fold_change'].apply(lambda x: 1 if x <= -5 else (1 if x >= 5 else 0))


true_labels = gpcr_vespa['foldchange_group']
predicted_labels = gpcr_vespa['VESPAl_group']

# Calculate confusion matrix
tn, fp, fn, tp = confusion_matrix(true_labels, predicted_labels).ravel()

# Calculate accuracy
accuracy = accuracy_score(true_labels, predicted_labels)

# Calculate specificity and sensitivity
specificity = tn / (tn + fp)
sensitivity = tp / (tp + fn)


# Calculate precision, recall and F1-score
precision = precision_score(true_labels, predicted_labels)
recall = recall_score(true_labels, predicted_labels)
f1 = f1_score(true_labels, predicted_labels)

# calculate FPR and TPR
fpr, tpr, thresholds = roc_curve(true_labels, predicted_labels)

# calculate AUCROC
auc_roc = auc(fpr, tpr)



print("Specificity:", specificity)
print("Sensitivity:", sensitivity)
print("Accuracy:", accuracy)
print("Precision:", precision)
print("F1-score:", f1)
print("AUCROC score: {:.3f}".format(auc_roc))

# Generate classification report
report = classification_report(true_labels, predicted_labels, labels=[0, 1], digits=4,zero_division="warn")

# Print report
print(report)
