==> Genebass_VEPs/3-1_from_Genebass_to_Dynamut2.ipynb <==
import pandas as pd
import numpy as np
import json

# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
gene_list_raw.rename(columns={'gene':'SYMBOL'},inplace=True)
gene_list_raw
# uni_list = gene_list_raw['Uniprot'].tolist()


VEPwithGB_icd10 = pd.read_csv('VEPwithGB_icd10_0323.csv', index_col = None)
VEPwithGB_icd10

VEPwithGB_icd10['mutation_from'] = VEPwithGB_icd10['Amino_acids'].str.split('/',expand=True)[0]
VEPwithGB_icd10['mutation_to'] = VEPwithGB_icd10['Amino_acids'].str.split('/',expand=True)[1]
VEPwithGB_icd10

np.unique(VEPwithGB_icd10['SYMBOL'])

VEPwithGB_icd10_uniid = pd.merge(VEPwithGB_icd10,gene_list_raw[['protein','SYMBOL','Uniprot']],on='SYMBOL',how='left')
VEPwithGB_icd10_uniid

np.unique(VEPwithGB_icd10_uniid['Uniprot'])

## make mutation list

unq_VEPwithGB_icd10 = VEPwithGB_icd10_uniid.drop_duplicates(subset=['SYMBOL','Protein_position','mutation_from','mutation_to'], keep='first')
unq_VEPwithGB_icd10 = unq_VEPwithGB_icd10.reset_index()
unq_VEPwithGB_icd10


for i in range(unq_VEPwithGB_icd10.shape[0]):
    # get the uniprot id 
    uniprot_id = unq_VEPwithGB_icd10['Uniprot'][i]

    mut_info = unq_VEPwithGB_icd10['mutation_from'][i]+str(unq_VEPwithGB_icd10['Protein_position'][i])+unq_VEPwithGB_icd10['mutation_to'][i]
    with open(f"MutationLists/{uniprot_id}.txt", 'a') as file:
        # write a new line to the file
        file.write("A "+mut_info+"\n")
    

unq_uniid_list = np.unique(unq_VEPwithGB_icd10['Uniprot'].tolist())
unq_uniid_list

# generate the bash file
for i in range(len(unq_uniid_list)):
        uniprot_id = unq_uniid_list[i]
        with open("curl_post.sh", 'a') as curl:
                # write a new line to the file
                curl.write(f"curl https://biosig.lab.uq.edu.au/dynamut2/api/prediction_list -X POST -i -F pdb_file=@/Users/joannahench/Documents/UCPH/MSc_Thesis/Genebass_VEPs/AF_pdb/{uniprot_id}.pdb -F mutations_list=@/Users/joannahench/Documents/UCPH/MSc_Thesis/Genebass_VEPs/MutationLists/{uniprot_id}.txt >>Dynamut2_out.txt"+"\n")    

## process the output file to get the job id, also handle other issues

# if runs successfully, we store the job id, if not successfullt(if the mutation position not match the pdb structure), record the protein index in the protein list
import json
job_ids = []
block_ids = []

with open('Dynamut2_out.txt') as f:
    block_idx = 0
    for line in f:
        if line.startswith('{'):
            data = json.loads(line.strip())
            # if successful, store the job id for further getting the results
            if 'job_id' in data:
                # print(f"Job ID: {data['job_id']}")
                job_ids.append(data['job_id'])
            # if met error, store the index for further delete the protein(these protein will not be used in the further analysis)
            elif 'error' in data:
                # print(f"Error: {data['error']}")
                # print(f"Block Index: {block_idx}")
                block_ids.append(block_idx)
            block_idx += 1

block_ids

[1, 8, 16, 17, 18, 25, 31, 33, 37, 38, 41, 50, 59, 61, 64]

unq_uniid_list

# Remove values accordding to indices
ok_uniid = np.delete(unq_uniid_list, block_ids)
ok_uniid

len(ok_uniid)

# generate a df showing the uniprot id and the job id
uni_job_id = pd.DataFrame({"Uniprot":ok_uniid,"Job_id":job_ids})
uni_job_id

# according to job_ids, generate the bash file curl_get.sh
for i in range(len(job_ids)):
        job_id = job_ids[i]
        with open("curl_get.sh", 'a') as curl:
                # write a new line to the file
                curl.write(f"curl https://biosig.lab.uq.edu.au/dynamut2/api/prediction_list -X GET -F job_id={job_id} >>res0323.txt"+"\n")    

## processing prediction

id = 0
data_df = pd.DataFrame()
with open('res0323.txt', 'r') as f:
    for line in f:
        data = json.loads(line.strip())
        df = pd.DataFrame(data).transpose().iloc[:-1]
        df[['mutation_from', 'Protein_position', 'mutation_to']] = df['mutation'].str.extract(r'(\D)(\d+)(\D)')
        df['Uniprot'] = ok_uniid[id] 
        data_df = pd.concat([data_df, df], ignore_index=True) 
        id = id + 1



# define the new column order
new_col_order = ['mutation_from', 'Protein_position', 'mutation_to','mutation','chain','prediction','Uniprot']
# reindex the columns with the new order
data_df = data_df.reindex(columns=new_col_order)
# merge with protein name
data_df = pd.merge(data_df,gene_list_raw.iloc[:,3:],on='Uniprot')
data_df

# output the results into a csv file
data_df.to_csv("Genebass_dynamut2_out.csv",index=False)
