==> Genebass_VEPs/4-1_Genebass_FoldX_preparation.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.metrics import r2_score
import re
import requests
import json
import os


# read in all gene names
gene_list_raw =pd.read_csv("230323_EST_ENSG_GENE_new.csv")
gene_list_raw = gene_list_raw.rename(columns={'gene':'SYMBOL'})
gene_list_raw

VEPwithGB_icd10 = pd.read_csv('VEPwithGB_icd10_0323.csv', index_col = None)
VEPwithGB_icd10['mutation_from'] = VEPwithGB_icd10['Amino_acids'].str.split('/',expand=True)[0]
VEPwithGB_icd10['mutation_to'] = VEPwithGB_icd10['Amino_acids'].str.split('/',expand=True)[1]
VEPwithGB_icd10_uniid = pd.merge(VEPwithGB_icd10,gene_list_raw[['protein','SYMBOL','Uniprot']],on='SYMBOL',how='left')
unq_VEPwithGB_icd10 = VEPwithGB_icd10_uniid.drop_duplicates(subset=['SYMBOL','Protein_position','mutation_from','mutation_to'], keep='first')
unq_VEPwithGB_icd10 = unq_VEPwithGB_icd10.reset_index()
unq_VEPwithGB_icd10

uniid_list = np.unique(unq_VEPwithGB_icd10['Uniprot'])
uniid_list

# for each GPCR, get AF pdb file

# a function to get the alphafold structure (pdb file) for each gene
def getAF(uniprot_id):
    # Set the API endpoint and the UniProt ID of the protein
    endpoint = "https://alphafold.ebi.ac.uk/files"

    # Send a GET request to the API to retrieve the protein's structure
    response = requests.get(f"{endpoint}/AF-{uniprot_id}-F1-model_v4.pdb")

    # Check the status code of the response to ensure it was successful
    if response.status_code == 200:
        filename = f"/Users/joannahench/Downloads/foldx5MacStd.tar_/Genebass/{uniprot_id}/{uniprot_id}.pdb"
        os.makedirs(os.path.dirname(filename), exist_ok=True)
        # Write the contents of the response to a PDB file
        with open(filename, "w") as f:
            f.write(response.text)
    else:
        print(f"Failed to retrieve structure for {uniprot_id}: {response.status_code}")



#get the structure
for i in range(len(uniid_list)):
    getAF(uniid_list[i])

# make mutation list


for i in range(unq_VEPwithGB_icd10.shape[0]):
    # get the uniprot id according to entryname and convert the Series to a string with index excluded
    uniprot_id = unq_VEPwithGB_icd10['Uniprot'][i]
    # gene = gene_list_raw.loc[gene_list_raw['protein']==gpcr_valid_missense.iloc[i,4],'gene'].to_string(index=False)
    # uniid_list.append(uniprot_id)
    # gene_list.append(gene)
    mut_info = unq_VEPwithGB_icd10.iloc[i,64]+'A'+str(unq_VEPwithGB_icd10.iloc[i,3])+unq_VEPwithGB_icd10.iloc[i,65]
    with open(f"/Users/joannahench/Downloads/foldx5MacStd.tar_/Genebass/{uniprot_id}/individual_list_{uniprot_id}.txt", 'a') as file:
        # write a new line to the file
        file.write(mut_info+';'+"\n")
    

# make bash file

# generate the bash file
for i in range(len(uniid_list)):
        uniprot_id = uniid_list[i]
        with open("foldx_command.sh", 'a') as foldx:
                foldx.write(f'cd {uniprot_id}'+"\n")
                # write the command
                foldx.write(f"../../foldx --command=BuildModel --pdb={uniprot_id}.pdb --mutant-file=individual_list_{uniprot_id}.txt"+"\n")
                foldx.write('cd ..'+"\n") 

../../foldx --help --command=BuildModel

../../foldx --command=BuildModel --pdb=/Users/joannahench/Downloads/foldx5MacStd.tar_/GPCRdb/P0DMS8/P0DMS8.pdb --mutant-file=/Users/joannahench/Downloads/foldx5MacStd.tar_/GPCRdb/P0DMS8/individual_list_P0DMS8.txt

nohup ./foldx_test.sh &
